refactor(sentry): sentry_init.pyを責務別3ファイルに分割
【背景】utils/sentry_init.pyが1000行超に肥大化し、Sentryイベント処理・
プリミティブ値変換・機微値スクラブの責務が単一ファイルに混在していた。

【変更内容】sentry_scrub_primitives.py（プリミティブ変換）、
sentry_scrub_values.py（値スクラブ）、sentry_scrub_events.py
（イベント処理）の3ファイルへ分割し、sentry_init.pyは初期化ロジック
のみを保持。対応する単体テストも責務別ファイルへ追随させ、
http_helpers.pyのimportパスを更新。

【影響範囲】公開APIの挙動・エクスポート構造は不変。内部モジュール
分割のみで、外部呼び出し側への破壊的変更なし。

品質ゲート: 1443 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
